integration ELO v1.2

+++ wsmatches/app/hspublic/elo.go
@@ -9,7 +9,7 @@ import (
 )
 
 // Имитация базы данных рейтингов
-var ratings = map[string]int{}
+var ratings = map[string]float64{}
 
 func EloLogic(match []models.Match, corps []models.Corporation) {
 	for _, c := range corps {
@@ -21,7 +21,7 @@ func EloLogic(match []models.Match, corps []models.Corporation) {
 	var elo []models.Corporation
 	for _, c := range corps {
 		el := c
-		el.Elo = ratings[c.Id]
+		el.Elo = int(ratings[c.Id])
 		elo = append(elo, el)
 	}
 
@@ -34,17 +34,7 @@ func EloLogic(match []models.Match, corps []models.Corporation) {
 	fmt.Println("файл сохранен " + path)
 }
 
-// Функция для расчета нового рейтинга
-//func calculateElo(winnerRating, loserRating int) (int, int) {
-//	const K = 32
-//	expectedScoreWinner := 1.0 / (1 + math.Pow(10, float64(loserRating-winnerRating)/400))
-//	expectedScoreLoser := 1 - expectedScoreWinner
-//	newWinnerRating := winnerRating + int(float64(K)*(1-expectedScoreWinner))
-//	newLoserRating := loserRating + int(float64(K)*(0-expectedScoreLoser))
-//	return newWinnerRating, newLoserRating
-//}
-
-func calculateEloRating(ratingA, ratingB, actualScoreA, actualScoreB int, kFactor int) (int, int) {
+func calculateEloRating(ratingA, ratingB, actualScoreA, actualScoreB float64, kFactor float64) (float64, float64) {
 	var scoreA, scoreB float64
 	if actualScoreA > actualScoreB {
 		// Победа игрока A
@@ -63,8 +53,8 @@ func calculateEloRating(ratingA, ratingB, actualScoreA, actualScoreB int, kFacto
 	expectedScoreA := 1 / (1 + math.Pow(10, float64(ratingB-ratingA)/400))
 	expectedScoreB := 1 / (1 + math.Pow(10, float64(ratingA-ratingB)/400))
 
-	newRatingA := int(float64(ratingA) + float64(kFactor)*(scoreA-expectedScoreA))
-	newRatingB := int(float64(ratingB) + float64(kFactor)*(scoreB-expectedScoreB))
+	newRatingA := ratingA + kFactor*(scoreA-expectedScoreA)
+	newRatingB := ratingB + kFactor*(scoreB-expectedScoreB)
 
 	return newRatingA, newRatingB
 }
@@ -73,7 +63,7 @@ func processMatches(matches []models.Match) {
 		corp1Rating := ratings[match.Corporation1Id]
 		corp2Rating := ratings[match.Corporation2Id]
 
-		newCorp1Rating, newCorp2Rating := calculateEloRating(corp1Rating, corp2Rating, match.Corporation1Score, match.Corporation2Score, 30)
+		newCorp1Rating, newCorp2Rating := calculateEloRating(corp1Rating, corp2Rating, float64(match.Corporation1Score), float64(match.Corporation2Score), 30)
 		ratings[match.Corporation1Id] = newCorp1Rating
 		ratings[match.Corporation2Id] = newCorp2Rating
 	}
